fix: resolve all ESLint errors and most warnings
- Separated form contexts and hooks into form-context.ts
- Fixed missing dependencies with useCallback in NutzapCard and CashuWallet
- Removed all ESLint errors (0 errors)
- Reduced warnings from 12 to 3

Remaining warnings:
- 2 fast refresh warnings in navigation-menu.tsx and sidebar.tsx
- 1 missing dependency warning in CashuWallet

Project continues to build successfully with 0 TypeScript errors and 0 vulnerabilities.
